% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_desilva.R
\name{GENLIST_r}
\alias{GENLIST_r}
\title{Generate the full list of polyploid genotypes}
\usage{
GENLIST_r(ng, na1, m2)
}
\arguments{
\item{ng}{Integer; total number of genotypes (must equal
\code{choose(na1 + m2 - 1, m2)}).}

\item{na1}{Integer; number of alleles including the null allele.}

\item{m2}{Integer; ploidy level.}
}
\value{
An integer matrix of dimensions \code{ng} x \code{m2}.  Each row
  is one genotype: a non-decreasing sequence of allele indices in
  \code{1:na1}.
}
\description{
Translates the GENLIST subroutine from de Silva et al. (2005).  Enumerates
all \code{ng} possible genotypes for a polyploid of ploidy \code{m2} with
\code{na1} alleles (including null) as an integer matrix whose rows are
sorted allele-index vectors.
}
\details{
Genotypes are combinations-with-replacement of \code{1:na1} taken \code{m2}
at a time, listed in lexicographic (non-decreasing) order.  The total
number of genotypes is:
\deqn{ng = \binom{na1 + m2 - 1}{m2} = G(m2 - 1,\, na1)}
}
\examples{
# All tetraploid genotypes with 2 alleles + null (na1 = 3)
GENLIST_r(ng = 15L, na1 = 3L, m2 = 4L)

}
\seealso{
\code{\link{INDEXG_r}}, \code{\link{RANMUL_r}}
}
