% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfreq-package.R
\docType{package}
\name{polyfreq-package}
\alias{polyfreq-package}
\alias{polyfreq}
\title{polyfreq: Fast De Silva Allele Frequency Estimation for Polyploids}
\description{
Provides an optimized implementation of the de Silva et al. (2005) EM
algorithm for estimating allele frequencies in polyploid populations.
Pure-R vectorized code is combined with an optional Rcpp + OpenMP back-end
for the computationally dominant sparse selfing matrix step.
}
\author{
\strong{Maintainer}: Author Name \email{author@example.com}

}
