% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_desilva.R
\name{SELFMAT_r}
\alias{SELFMAT_r}
\title{Build the polyploid selfing matrix}
\usage{
SELFMAT_r(ng, na1, ag, m2)
}
\arguments{
\item{ng}{Integer; number of genotypes.}

\item{na1}{Integer; number of alleles including the null allele.}

\item{ag}{Integer matrix (\code{ng} x \code{m2}) of all genotypes, as
returned by \code{\link{GENLIST_r}}.}

\item{m2}{Integer; ploidy level (must be even; \code{m = m2 / 2}).}
}
\value{
An integer matrix of dimensions \code{ng} x \code{ng}.
}
\description{
Translates the SELFMAT subroutine from de Silva et al. (2005).  Constructs
the \code{ng} x \code{ng} integer selfing matrix \code{smat}, where entry
\code{smat[g1, g2]} counts the number of gamete-pair combinations from
selfing parent \code{g1} that produce offspring genotype \code{g2}.
}
\details{
The matrix is divided by \code{(G_r(m - 1, m + 1))^2} (= \code{choose(m2,
m)^2}) inside \code{de_silva_freq_EM} to convert counts to proportions
(this divisor is precomputed as \code{smatdiv}).
}
\seealso{
\code{\link{GENLIST_r}}, \code{\link{RANMUL_r}},
  \code{\link{INDEXG_r}}
}
